Move the entries and list widgets to their respective packages
