+++ src/main/java/fr/aeldit/ctms/gui/GroupsScreen.java
@@ -1,27 +1,21 @@
 package fr.aeldit.ctms.gui;
 
+import fr.aeldit.ctms.gui.widgets.GroupsListWidget;
 import fr.aeldit.ctms.textures.CTMSelector;
 import fr.aeldit.ctms.textures.Group;
 import fr.aeldit.ctms.textures.entryTypes.CTMPack;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.gui.Element;
-import net.minecraft.client.gui.Selectable;
 import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.tooltip.Tooltip;
-import net.minecraft.client.gui.widget.*;
+import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.screen.ScreenTexts;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
-import org.apache.commons.compress.utils.Lists;
-import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.List;
 import java.util.Objects;
 
 import static fr.aeldit.ctms.Utils.TEXTURES_HANDLING;
@@ -107,110 +101,5 @@ public class GroupsScreen extends Screen
                         .dimensions(width / 2 - 100, height - 26, 200, 20)
                         .build()
         );
-    }
-
-    /**
-     * Modified by me to fit my purpose
-     *
-     * @author dicedpixels (<a href="https://github.com/dicedpixels">...</a>)
-     */
-    private static class GroupsListWidget extends ElementListWidget<GroupEntry>
-    {
-        private final GroupEntryBuilder builder = new GroupEntryBuilder(client, width);
-
-        //? if <1.20.4 {
-        /*public GroupsListWidget(
-                MinecraftClient client, int width, int height, int top, int bottom, int itemHeight,
-                CTMSelector ctmSelector
-        )
-        {
-            super(client, width, height, top, bottom, itemHeight);
-            this.ctmSelector = ctmSelector;
-        }
-
-        *///?} else {
-        public GroupsListWidget(
-                MinecraftClient client, int width, int height, int y, int itemHeight
-        )
-        {
-            super(client, width, height, y, itemHeight);
-        }
-        //?}
-
-        public void add(Group group)
-        {
-            addEntry(builder.build(group));
-        }
-    }
-
-    private record GroupEntryBuilder(MinecraftClient client, int width)
-    {
-        @Contract("_ -> new")
-        public @NotNull GroupsScreen.GroupEntry build(@NotNull Group group)
-        {
-            var layout = DirectionalLayoutWidget.horizontal().spacing(5);
-            var text = new TextWidget(
-                    160, 20 + 2,
-                    Text.of(group.isEnabled()
-                            ? group.getGroupName()
-                            : Formatting.RED + Text.of(Formatting.ITALIC + group.getGroupName()).getString()
-                    ),
-                    client.textRenderer
-            );
-            var toggleButton = CyclingButtonWidget.onOffBuilder()
-                    .omitKeyText()
-                    .initially(group.isEnabled())
-                    .build(0, 0, 30, 20, Text.empty(),
-                           (button, value) -> group.toggle()
-                    );
-            toggleButton.setTooltip(Tooltip.of(group.getButtonTooltip()));
-            text.alignLeft();
-            layout.add(EmptyWidget.ofWidth(15));
-            layout.add(text);
-            layout.add(toggleButton);
-            layout.refreshPositions();
-            layout.setX(width / 2 - layout.getWidth() / 2);
-            return new GroupEntry(group, layout);
-        }
-    }
-
-    static class GroupEntry extends ElementListWidget.Entry<GroupEntry>
-    {
-        private final Group group;
-        private final LayoutWidget layout;
-        private final List<ClickableWidget> children = Lists.newArrayList();
-
-        GroupEntry(Group group, LayoutWidget layout)
-        {
-            this.group = group;
-            this.layout = layout;
-            this.layout.forEachChild(this.children::add);
-        }
-
-        @Override
-        public List<? extends Selectable> selectableChildren()
-        {
-            return children;
-        }
-
-        @Override
-        public List<? extends Element> children()
-        {
-            return children;
-        }
-
-        @Override
-        public void render(
-                @NotNull DrawContext context, int index, int y, int x,
-                int entryWidth, int entryHeight, int mouseX, int mouseY,
-                boolean hovered, float delta
-        )
-        {
-            context.drawTexture(group.getIdentifier(), x, y + 2, 0, 0, 16, 16, 16, 16);
-            layout.forEachChild(child -> {
-                child.setY(y);
-                child.render(context, mouseX, mouseY, delta);
-            });
-        }
     }
 }
