+++ src/main/java/fr/aeldit/ctms/gui/ScreenUtils.java
@@ -6,5 +6,7 @@ import net.minecraft.text.Text;
 public class ScreenUtils
 {
     public static final Text TEXT_RESET = Text.translatable("ctms.screen.config.reset");
+    public static final Text TEXT_OPEN = Text.translatable("ctms.screen.open");
     public static final Tooltip TOOLTIP_RESET = Tooltip.of(Text.translatable("ctms.screen.config.reset.tooltip"));
+    public static final Tooltip TOOLTIP_OPEN = Tooltip.of(Text.translatable("ctms.screen.open.tooltip"));
 }
